refactor(core): mapper 위치 이동
