+++ member-service/src/main/java/com/example/memberservice/common/security/handler/OAuthLoginFailureHandler.java
@@ -1,7 +1,7 @@
 package com.example.memberservice.common.security.handler;
 
 
-import static com.example.memberservice.common.exception.ResponseDtoMapper.getErrorResponse;
+import static com.example.memberservice.common.web.ResponseDtoMapper.getErrorResponse;
 
 import com.example.memberservice.common.exception.ErrorCode;
 import org.hexagon.core.dto.Empty;
